chore: split orders-http-state tests out of orders-http.integration.test.ts
The issue-70 order-detail test pushed orders-http.integration.test.ts
over the 400-line eslint max-lines cap. Split out the state-change
(#25) tests into their own file with their own boot() copy, same
pattern as catalog-http.integration.test.ts / catalog-http-ingest.
integration.test.ts.

Follow-up to PR 69 (issue 67) - issue 70.
